pyproject.toml
[tool.poetry]
name = "fast-api-todo"
version = "0.1.0"
description = ""
authors = ["Valery Shcherbakov <val032606@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.111.0"}
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
pydantic-settings = "^2.3.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
databases = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
